feat(server): add new interfaces file
